+++ CHANGELOG.md
@@ -11,6 +11,7 @@ n./a.
 * Implement a `VisualTransformation` to handle grouping- and decimal-separators for better user experience and reusability.
 * Fix a bug where the text input to enter the hours worked was not visible sometimes.
 * `TransferScreen` shows the name of the type for which a transfer is being created or edited.
+* Move database and repository declaration from `MainActivity` to `ChaChingApplication` to make them accessible from Android widgets.
 
 <br/>